[#15] Get keywords with order_by and fix comments
