+++ lib/crawler/keyword/queries/keyword_query.ex
@@ -3,8 +3,12 @@ defmodule Crawler.Keyword.Queries.KeywordQuery do
 
   alias Crawler.Keyword.Schemas.Keyword
 
-  def user_keywords_query(user_id), do: where(Keyword, [keyword], keyword.user_id == ^user_id)
+  def user_keywords_query(user_id),
+    do:
+      Keyword
+      |> where([k], k.user_id == ^user_id)
+      |> order_by(desc: :inserted_at)
 
   def user_keywords_query(name, user_id),
-    do: where(Keyword, [keyword], keyword.user_id == ^user_id and keyword.name == ^name)
+    do: where(Keyword, [k], k.user_id == ^user_id and k.name == ^name)
 end
