+++ lib/crawler_web/templates/home/index.html.heex
@@ -1,4 +1,4 @@
-<nav class="mt-3" aria-label="breadcrumb">
+<div class="mt-3" aria-label="breadcrumb">
   <div class="container">
     <h1 aria-current="page">
       <%= gettext("Home") %>
@@ -8,7 +8,7 @@
       <%= gettext("Please choose a CSV file:") %>
     </h4>
 
-    <%= form_for @changeset, Routes.home_path(@conn, :create), [method: "post" , class: "row pt-2" , multipart: true], fn
+    <%= form_for @changeset, Routes.keyword_path(@conn, :create), [method: "post" , class: "row pt-2" , multipart: true], fn
   f-> %>
       <div class="col-3">
         <%= file_input(f, :file, class: "form-control", required: true, accept: 'text/csv') %>
@@ -45,4 +45,4 @@
       </div>
     <% end %>
   </div>
-</nav>
+</div>
