+++ lib/crawler_web/views/keyword_view.ex
@@ -1,0 +1,3 @@
+defmodule CrawlerWeb.KeywordView do
+  use CrawlerWeb, :view
+end
